requirements.txt
numpy==1.24.4
pyro_ppl==1.9.0
scikit_learn==1.2.0
scipy==1.12.0
torch_geometric==2.5.2
torch_scatter==2.1.2
tqdm==4.66.1
transformers==4.39.1
Unidecode==1.3.8
pypinyin
sentencepiece
sacremoses